* Application/FSController.m         * Application/FSDocument.m         * Application/FSFunctionHelp.m         * Application/FSTableController+Formula.m         * Application/FSTableController.m         * Application/Inspection/FSCellInspectorPane.m         * Application/Inspection/FSDocumentInspectorPane.m         * Application/Inspection/FSInspection.m         * Application/Inspection/FSTableInspectorPane.m         * Application/Inspection/FSViewInspectorPane.m         * Application/UI/FSFormulaTable.m         * Application/UI/FSMatrix.m         * Application/UI/FSTableView+Keyboard.m         * Application/UI/FSTableView+Layout.m         * Application/UI/ImageAndTextCell.m         * FSCore/FSKeyGroup.m         * FSCore/FSKeySet.h         * FSCore/FSKeySet.m         * FSCore/FSTable.m         * FSCore/FSVariable.m         Some NS(U)Integer and float -> CGFloat transitions and         other minor fixes
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2630 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/FlexiSheet/Application/FSFunctionHelp.m
@@ -5,7 +5,7 @@
 //  Created by Stefan Leuker on 17-OCT-2001.
 //  Copyright (c) 2001-2003 Stefan Leuker. All rights reserved.
 //
-//  $Id: FSFunctionHelp.m,v 1.1 2008/10/14 15:03:45 hns Exp $
+//  $Id: FSFunctionHelp.m,v 1.2 2014/01/26 09:23:52 buzzdee Exp $
 
 #import "FSFunctionHelp.h"
 #import "FSFunction.h"
@@ -68,7 +68,7 @@
     }
 }
 
-- (int)browser:(NSBrowser *)sender numberOfRowsInColumn:(int)column
+- (NSInteger)browser:(NSBrowser *)sender numberOfRowsInColumn:(NSInteger)column
 {
     if (sender == groups) {
         return [[FSFunction allGroupNames] count]+1;
@@ -79,7 +79,7 @@
     return 0;
 }
 
-- (void)browser:(NSBrowser *)sender willDisplayCell:(id)cell atRow:(int)row column:(int)column
+- (void)browser:(NSBrowser *)sender willDisplayCell:(id)cell atRow:(NSInteger)row column:(NSInteger)column
 {
     [cell setLeaf:YES];
     if (sender == groups) {
